requirements.txt
pydantic~=2.10.6
openai~=1.66.3
tenacity~=9.0.0
pyyaml~=6.0.2
loguru~=0.7.3
numpy
datasets<=3.1.0
fastapi~=0.115.11
tiktoken~=0.9.0
RestrictedPython
html2text~=2024.2.26
gymnasium~=1.0.0
pillow~=10.4.0
browsergym~=0.13.3
uvicorn~=0.34.0
unidiff~=0.7.5
browser-use~=0.1.40
googlesearch-python~=1.3.0
baidusearch~=1.0.3
duckduckgo-search~=7.4.0

aiofiles~=24.1.0
pydantic_core~=2.27.2
colorama~=0.4.6
playwright~=1.50.0
diskcache~=5.6.3  # For persistent LLM response caching

# Enhanced academic search
requests>=2.32.2
beautifulsoup4~=4.12.3
lxml>=5.3.0
scholarly~=1.7.11  # For Google Scholar access
